Fix asset naming
In production all assets are placed into a single directory. Therefore
the webpack-assets-manifest plugin does not include any path for asset
keys, e.g. the imported asset `images/favicon.ico` is only emitted as
`favicon.ico` but the Rails application expects the used import path of
`images/favicon.ico`.

This CL changes the customization to lookup the child module for a given
asset and derives the original name from the module name.