feat: Show inline nested traces
This provides a better overview on what things actually are happening in
all services due to a single request. When a spans triggers exactly one
new traces (usually from another service), that trace will be directly
rendered inline the graph view.

The graph now has a color palette, and will color the bars differently
based on a hash on the application UUID. This eases identifing which
spans belong to which application.

The implementation has been changed to also store a span with the same
details as the trace, as the graph now only needs to fetch spans at all.
The top row won't be a manually added trace anymore. This also prepares
the application to not store so much data for traces anymore, but really
search on all spans.