fix: Remove unnessary breaking database migration
The activity_id database migration breaks on exists data because it
tries to add a non-NULL column. It also is not necessary at all anymore,
as the current nested-ordered algorithm loading spans does not use that
column anymore. The activity UUID itself can also be looked up via the
trace association.