feat: Improve performance by preload nested spans
The span decorator didn't reload the associations needed for serializing
the nested spans yet. This CL adds the same preloading statements to the
span decorator as already used by the trace decorator. This greatly cuts
down the number of requests. They are still not all inlined, as we still
need one call per actual nested span.